fix decimal with precision

+++ ft_processor.c
@@ -6,7 +6,7 @@
 /*   By: mrosie <mrosie@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/19 12:38:50 by mrosie            #+#    #+#             */
-/*   Updated: 2020/11/24 19:03:24 by mrosie           ###   ########.fr       */
+/*   Updated: 2020/11/25 13:27:52 by mrosie           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,11 +25,11 @@ int	ft_processor(char type, t_spec *sp, va_list ar)
 	else if (type == 'u')
 		return (ft_print_number(*sp, va_arg(ar, unsigned int), 10, 0));
 	else if (type == 'x')
-		return (ft_print_number(*sp, va_arg(ar, unsigned long), 16, 0));
+		return (ft_print_number(*sp, va_arg(ar, unsigned int), 16, 0));
 	else if (type == 'X')
-		return (ft_print_number(*sp, va_arg(ar, unsigned long), 16, 1));
+		return (ft_print_number(*sp, va_arg(ar, unsigned int), 16, 1));
 	else if (type == 'o')
-		return (ft_print_number(*sp, va_arg(ar, unsigned long), 8, 0));
+		return (ft_print_number(*sp, va_arg(ar, unsigned int), 8, 0));
 	else if (type == '%')
 		return (ft_print_char(*sp, '%'));
 	return (-1);
